tests: clean go1.24 build tags
