--- pkg/golinters/usetesting/testdata/usetesting_go124_configuration.yml
@@ -1,12 +1,0 @@
-version: "2"
-
-linters:
-  settings:
-    usetesting:
-      os-create-temp: false
-      os-mkdir-temp: false
-      os-setenv: true
-      os-temp-dir: true
-      os-chdir: false
-      context-background: false
-      context-todo: false
